N/A #in-progress #time 30m #comment Refactor WorkshopCard and abstract styles in .js file
