+++ src/App.js
@@ -5,9 +5,9 @@ import Home from './components/Pages/Home/Home';
 import Nav from './components/Nav';
 import AllWorkshops from './components/Pages/AllWorkshops/AllWorkshops';
 import WorkshopOverview from './components/Pages/workshopOverview/WorkshopOverview';
-import WorkshopFeedback from './components/Pages/WorkshopFeedback';
-import WorkshopWorksheet from './components/Pages/WorkshopWorksheet';
-import WorkshopContent from './components/Pages/WorkshopContent';
+// import WorkshopFeedback from './components/Pages/WorkshopFeedback';
+// import WorkshopWorksheet from './components/Pages/WorkshopWorksheet';
+// import WorkshopContent from './components/Pages/WorkshopContent';
 
 import './App.css';
 
@@ -25,9 +25,9 @@ function App() {
 							</Route>
 							<Route exact path="/workshops" component={AllWorkshops} />
 							<Route exact path="/workshop/overview/:ID" component={WorkshopOverview} />
-							<Route exact path="/workshop/feedback/:ID" component={WorkshopFeedback} />
+							{/* <Route exact path="/workshop/feedback/:ID" component={WorkshopFeedback} />
 							<Route exact path="/workshop/content/:ID" component={WorkshopContent} />
-							<Route exact path="/workshop/worksheet/:ID" component={WorkshopWorksheet} />
+							<Route exact path="/workshop/worksheet/:ID" component={WorkshopWorksheet} /> */}
 						</Switch>
 					</main>
 				</div>
