+++ src/components/Pages/AllWorkshops/style.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+export const Main = styled.section`
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	justify-content: center;
+	margin-top: 4rem;
+`;
+
+export const otherStyle = null;
