+++ src/components/WorkshopCard/WorkshopCard.stories.js
@@ -1,0 +1,10 @@
+import WorkshopCard from './WorkshopCard';
+
+export default {
+	component: WorkshopCard,
+	title: 'WorkshopCard',
+};
+
+export const defaultCard = () => {
+	WorkshopCard();
+};
