--- src/components/workshopCard.js
@@ -1,90 +1,0 @@
-// import React from "react";
-// import Card from "@material-ui/core/Card";
-// import CardMedia from "@material-ui/core/CardMedia";
-// import CardContent from "@material-ui/core/CardContent";
-// import Typography from "@material-ui/core/Typography";
-// import styled from 'styled-components';
-
-// let imgUrl = "www.image.com"
-// const tagItems = [{
-//     title : "Covid-19"
-// },{
-//     title : "Radio"
-// },{
-//     title : "Arabic"}
-// ]
-
-// // const Tag =
-
-// export default function WorkshopCard() {
-// 	return(
-// 		<Card className="card">
-//         <CardMedia className="card__image" image={imgUrl}/>
-//         <List>{tagItems.map(item => (
-//             <ListItem button key={item.title} component={StyledLink} to={item.titleHref}>
-// 			<ListItemText primary={item.title} />
-// 			</ListItem>
-//         ))}
-//         </List>
-//         <Typography className="card__tag_R">
-
-//         </Typography>
-//     </Card>
-
-// 	)
-// }
-
-/*
-[
-	{
-	  "_table": {
-		"_base": {
-		  "_airtable": {
-			"requestTimeout": 300000
-		  },
-		  "_id": "appW2989WcgTxarh0"
-		},
-		"id": null,
-		"name": "workshops"
-	  },
-	  "id": "reclXZv3JtH77KpzW",
-	  "_rawJson": {
-		"id": "reclXZv3JtH77KpzW",
-		"fields": {
-		  "average_rating": 4,
-		  "title": "Workshop uno"
-		},
-		"createdTime": "2020-06-08T15:21:46.000Z"
-	  },
-	  "fields": {
-		"average_rating": 4,
-		"title": "Workshop uno"
-	  }
-	},
-	{
-	  "_table": {
-		"_base": {
-		  "_airtable": {
-			"requestTimeout": 300000
-		  },
-		  "_id": "appW2989WcgTxarh0"
-		},
-		"id": null,
-		"name": "workshops"
-	  },
-	  "id": "recIzSfHrfdbTk05n",
-	  "_rawJson": {
-		"id": "recIzSfHrfdbTk05n",
-		"fields": {
-		  "average_rating": 3,
-		  "title": "Workshop dos"
-		},
-		"createdTime": "2020-06-08T15:21:46.000Z"
-	  },
-	  "fields": {
-		"average_rating": 3,
-		"title": "Workshop dos"
-	  }
-	}
-  ]
-  */
