+++ src/styles/index.js
@@ -1,0 +1,2 @@
+// export * as colours from "./colours"
+export * as componentsStyles from './componentsStyles';
